requirements.txt
Django==1.10.3
django-filter==0.15.3
djangorestframework==3.5.2
Markdown==2.6.7
Pillow==3.4.2
